Rename package JJ to jackson Change name ExtractFromString to StringExtractor
