+++ src/main/java/com/solvd/sql/jackson/JSONUtils.java
@@ -1,9 +1,9 @@
-package com.solvd.sql.JJ;
+package com.solvd.sql.jackson;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.solvd.enums.Paths;
-import com.solvd.util.ExtractFromString;
+import com.solvd.util.StringExtractor;
 
 import java.io.File;
 import java.io.IOException;
@@ -22,7 +22,7 @@ public class JSONUtils {
     // Input assumes the class name is the first String. Name convention would be class name follow by a number
     public static Object readJSON(String filename) {
         try {
-            String className = ExtractFromString.extractClassName(filename);
+            String className = StringExtractor.extractClassName(filename);
             Class<?> objectClass = Class.forName(Paths.MODELFOLDER.getPath() + "." + className);
             ObjectMapper objectMapper = new ObjectMapper().enable(SerializationFeature.INDENT_OUTPUT);
             return objectMapper.readValue(new File(Paths.JSONFOLDER.getPath() + objectClass.getSimpleName() + ".json"), objectClass);
